Migrate link layer and IP tests from the main codebase into the tests package because they are dependent on the packet capture configuration
